requirements.txt
Appium-Python-Client~=4.4.0
selenium~=4.27.1
pydantic~=2.10.2
PySide6~=6.8.0.2
